Reorder: put git at the top
The git part of the lesson no longer depends on RStudio parts before it, so it can be put at the top. This solves inefficiency in flow, where a note told students to "go quickly through the first part and concentrate on the second one". Now it's more natural, with the important part being right there at the top.